+++ plugins/CustomVariables/Columns/Base.php
@@ -8,14 +8,16 @@
 
 namespace Piwik\Plugins\CustomVariables\Columns;
 
+use Piwik\Columns\DimensionSegmentFactory;
 use Piwik\Piwik;
 use Piwik\Plugin\Dimension\VisitDimension;
 use Piwik\Plugins\CustomVariables\Segment;
 use Piwik\Plugins\CustomVariables\CustomVariables;
+use Piwik\Segment\SegmentsList;
 
 class Base extends VisitDimension
 {
-    protected function configureSegmentsFor($segmentNameSuffix)
+    protected function configureSegmentsFor($segmentNameSuffix, SegmentsList $segmentsList, DimensionSegmentFactory $dimensionSegmentFactory)
     {
         $numCustomVariables = CustomVariables::getNumUsableCustomVariables();
 
@@ -24,14 +26,14 @@ class Base extends VisitDimension
         $segment->setSegment('customVariable' . $segmentNameSuffix);
         $segment->setName($this->getName() . ' (' . Piwik::translate('CustomVariables_ScopeVisit') . ')');
         $segment->setUnionOfSegments($this->getSegmentColumns('customVariable' . $segmentNameSuffix, $numCustomVariables));
-        $this->addSegment($segment);
+        $segmentsList->addSegment($dimensionSegmentFactory->createSegment($segment));
 
         $segment = new Segment();
         $segment->setType('dimension');
         $segment->setSegment('customVariablePage' . $segmentNameSuffix);
         $segment->setName($this->getName() . ' (' . Piwik::translate('CustomVariables_ScopePage') . ')');
         $segment->setUnionOfSegments($this->getSegmentColumns('customVariablePage' . $segmentNameSuffix, $numCustomVariables));
-        $this->addSegment($segment);
+        $segmentsList->addSegment($dimensionSegmentFactory->createSegment($segment));
 
         $segmentSuffix = 'v';
         if (strtolower($segmentNameSuffix) === 'name') {
@@ -44,14 +46,14 @@ class Base extends VisitDimension
             $segment->setSqlSegment('log_visit.custom_var_' . $segmentSuffix . $i);
             $segment->setName(Piwik::translate('CustomVariables_ColumnCustomVariable' . $segmentNameSuffix) . ' ' . $i
                     . ' (' . Piwik::translate('CustomVariables_ScopeVisit') . ')');
-            $this->addSegment($segment);
+            $segmentsList->addSegment($dimensionSegmentFactory->createSegment($segment));
 
             $segment = new Segment();
             $segment->setSegment('customVariablePage' . $segmentNameSuffix . $i);
             $segment->setSqlSegment('log_link_visit_action.custom_var_' . $segmentSuffix . $i);
             $segment->setName(Piwik::translate('CustomVariables_ColumnCustomVariable' . $segmentNameSuffix) . ' ' . $i
                     . ' (' . Piwik::translate('CustomVariables_ScopePage') . ')');
-            $this->addSegment($segment);
+            $segmentsList->addSegment($dimensionSegmentFactory->createSegment($segment));
         }
     }